requirements-dev.txt
ruff==0.6.8
mypy==1.11.2
mypy-extensions==1.0.0
types-requests==2.32.0.20240914
setuptools>=70.0.0
wheel>=0.43.0
twine>=5.1.0
python-dotenv>=1.0.1
twine>=5.1.1
